Updated code. dynamic table without shrink and open addressing not updated yet

+++ open_addressing.h
@@ -9,24 +9,24 @@
 struct bin {
 	bool is_free : 1;
 	bool is_deleted : 1;
-	uint32_t key;
+	unsigned int key;
 };
 
 struct hash_table {
     struct bin *table;
-    size_t size;
-    size_t used;
-    size_t active;
+    unsigned int size;
+    unsigned int used;
+    unsigned int active;
     double load_limit;
     // only used in primes code, but we share the header, so...
-    uint32_t primes_idx;
+    unsigned int primes_idx;
 };
 
-struct hash_table *empty_table(size_t size, double load_limit);
+struct hash_table *empty_table(unsigned int size, double load_limit);
 void delete_table(struct hash_table *table);
 
-void insert_key  (struct hash_table *table, uint32_t key);
-bool contains_key(struct hash_table *table, uint32_t key);
-void delete_key  (struct hash_table *table, uint32_t key);
+void insert_key  (struct hash_table *table, unsigned int key);
+bool contains_key(struct hash_table *table, unsigned int key);
+void delete_key  (struct hash_table *table, unsigned int key);
 
 #endif
